Changing how to access stuff
Former-commit-id: 211d5cb0ed284b4cc7e44e9c712f730569a3c98f

+++ Intersect Library/Intersect Library/Logging/LogLevel.cs
@@ -1,6 +1,6 @@
 ﻿namespace Intersect.Logging
 {
-    public enum LogLevel : int
+    public enum LogLevel : byte
     {
         Error,
         Warn,
